// src/app/models/afiliado.model.ts
export interface Afiliado {
    id?: string;
    aprobado: boolean;
    categoria: string;
    fotos: string[];
    img: string;
    latitud: number;
    longitud: number;
    nombre: string;
    puntos: number;
    rating: number;
    rfc: string;
    telefono: string;
    total: number;
    ubicacion: string;
}